+++ scripts/uploadSprite.ts
@@ -4,14 +4,21 @@ import fs from 'fs'
 import path from 'path'
 
 const version = process.argv[2]
-AWS.config.update({ region: 'us-west-2' })
+AWS.config.update({
+  region: 'us-west-2',
+  credentials: {
+    accessKeyId: process.env.AWS_ACCESS_KEY_ID,
+    secretAccessKey: process.env.AWS_SECRET_ACCESS_KEY
+  }
+})
 const s3 = new AWS.S3({ apiVersion: '2006-03-01' })
 
 const uploadParams: PutObjectRequest = {
   Bucket: 'cdn-icons-occ',
   Key: '',
   CacheControl: 'max-age=2592000',
-  ContentType: 'image/svg+xml'
+  ContentType: 'image/svg+xml',
+  ACL: 'bucket-owner-full-control'
 }
 
 const fileStream = fs.createReadStream(