+++ utils/icon-groups.ts
@@ -1,0 +1,379 @@
+interface IconGroup {
+  name: string
+  icons: string[]
+}
+export const iconGroups: IconGroup[] = [
+  {
+    name: 'Chevrons',
+    icons: ['chevron-down', 'chevron-up', 'chevron-left', 'chevron-right']
+  },
+  {
+    name: 'Arrows',
+    icons: [
+      'arrow-down',
+      'arrow-up',
+      'arrow-left',
+      'arrow-right',
+      'arrow-up-right',
+      'arrow-down-up',
+      'arrow-left-right',
+      'arrow-down-circle',
+      'arrow-up-circle',
+      'arrow-left-circle',
+      'arrow-right-circle',
+      'arrow-house-up'
+    ]
+  },
+  {
+    name: 'Box arrows',
+    icons: [
+      'box-arrow-up',
+      'box-arrow-down',
+      'box-arrow-left',
+      'box-arrow-right',
+      'box-arrow-up-left',
+      'box-arrow-up-right',
+      'box-arrow-down-right',
+      'box-arrow-down-left',
+      'box-arrow-in-up',
+      'box-arrow-in-down',
+      'box-arrow-in-left',
+      'box-arrow-in-right',
+      'box-arrow-in-up-left',
+      'box-arrow-in-up-right',
+      'box-arrow-in-down-right',
+      'box-arrow-in-down-left'
+    ]
+  },
+  {
+    name: 'Symbols',
+    icons: [
+      'dash',
+      'plus',
+      'slash',
+      'x',
+      'check',
+      'info',
+      'question',
+      'exclamation',
+      'double-check',
+      'dash-circle',
+      'plus-circle',
+      'slash-circle',
+      'x-circle',
+      'check-circle',
+      'info-circle',
+      'question-circle',
+      'exclamation-circle',
+      'alert'
+    ]
+  },
+  {
+    name: 'Controls',
+    icons: ['three-dots', 'three-dots-vertical']
+  },
+  {
+    name: 'Misc',
+    icons: [
+      'bookmarks',
+      'bookmark',
+      'share',
+      'download',
+      'upload',
+      'dot',
+      'fullscreen',
+      'fullscreen-exit',
+      'link-45-deg',
+      'view-stacked',
+      'view-list',
+      'lightning-charge',
+      'heart',
+      'heart-filled',
+      'location',
+      'chart-line-up',
+      'filter-alt',
+      'menu',
+      'chart',
+      'code',
+      'math',
+      'on-off',
+      'cards',
+      'copy',
+      'dark-light',
+      'go-to-direction',
+      'match-job',
+      'percent',
+      'pointer',
+      'refresh'
+    ]
+  },
+  {
+    name: 'Files folders',
+    icons: [
+      'files',
+      'files-alt',
+      'journals',
+      'folder',
+      'folder-symlink',
+      'folder-plus',
+      'folder-minus',
+      'folder-check',
+      'folder-x',
+      'folder-2',
+      'folder-2-open',
+      'journals-alt',
+      'image',
+      'images',
+      'card-image',
+      'card-text',
+      'card-heading',
+      'card-checklist',
+      'table',
+      'file-text',
+      'file-checklist',
+      'file-user'
+    ]
+  },
+  {
+    name: 'People',
+    icons: [
+      'person',
+      'people',
+      'person-add',
+      'person-dash',
+      'person-x',
+      'person-block',
+      'person-v-card',
+      'person-v-card-alt',
+      'person-add-circle',
+      'people-circle'
+    ]
+  },
+  {
+    name: 'Devices',
+    icons: [
+      'mobile',
+      'mobile-landscape',
+      'tablet',
+      'tablet-landscape',
+      'display',
+      'tv',
+      'laptop',
+      'printer',
+      'music-player',
+      'camera',
+      'watch',
+      'smartwatch',
+      'mobile-flip',
+      'robot',
+      'battery',
+      'battery-half',
+      'battery-full',
+      'battery-charging',
+      'calculator',
+      'mouse',
+      'headphones',
+      'controller',
+      'usb'
+    ]
+  },
+  {
+    name: 'Time',
+    icons: [
+      'clock',
+      'clock-history',
+      'stopwatch',
+      'alarm',
+      'calendar',
+      'calendar-date',
+      'calendar-event',
+      'calendar-week',
+      'calendar-range',
+      'calendar-plus',
+      'calendar-minus',
+      'calendar-x',
+      'calendar-check'
+    ]
+  },
+  {
+    name: 'Communication',
+    icons: [
+      'chat',
+      'chat-dots',
+      'chat-text',
+      'chat-quote',
+      'chat-smile',
+      'end-chat',
+      'bell',
+      'bell-slash',
+      'envelope',
+      'envelope-open',
+      'envelope-exclamation',
+      'globe',
+      'search',
+      'phone',
+      'send',
+      'qr-code',
+      'qr-code-scan'
+    ]
+  },
+  {
+    name: 'Weather',
+    icons: [
+      'stars',
+      'lightning',
+      'moon',
+      'sun',
+      'star-alt',
+      'star',
+      'star-filled',
+      'aistars'
+    ]
+  },
+  {
+    name: 'Security',
+    icons: ['lock', 'unlock', 'fingerprint', 'security-check', 'verified-check']
+  },
+  {
+    name: 'Commerce',
+    icons: [
+      'bag',
+      'bag-plus',
+      'bag-dash',
+      'bag-check',
+      'bag-x',
+      'cart',
+      'cart-plus',
+      'cart-dash',
+      'cart-check',
+      'cart-x',
+      'wallet',
+      'shop',
+      'receipt',
+      'upc',
+      'upc-scan',
+      'truck',
+      'bank',
+      'truck-alt',
+      'cash',
+      'cash-stack',
+      'dollar',
+      'dollar-alt',
+      'euro',
+      'bitcoin',
+      'exchange',
+      'coin',
+      'cash-coin',
+      'piggy-bank'
+    ]
+  },
+  {
+    name: 'Tools',
+    icons: [
+      'pencil',
+      'pen',
+      'gear',
+      'tools',
+      'brush',
+      'magnet',
+      'wrench',
+      'wrench-alt',
+      'pencil-alt'
+    ]
+  },
+  {
+    name: 'Sort and filter',
+    icons: [
+      'filter',
+      'filter-left',
+      'filter-right',
+      'sort-up',
+      'sort-down',
+      'sort-up-alt',
+      'sort-down-alt',
+      'sort-alpha-down',
+      'sort-alpha-up',
+      'sort-alpha-down-alt',
+      'sort-alpha-up-alt',
+      'sort-numeric-down',
+      'sort-numeric-up',
+      'sort-numeric-down-alt',
+      'sort-numeric-up-alt'
+    ]
+  },
+  {
+    name: 'Typography',
+    icons: ['list-ol', 'list-ul', 'type-bold', 'type-italic', 'type-underline']
+  },
+  {
+    name: 'Layout',
+    icons: ['columns', 'columns-gap', 'text-window', 'grid', 'add-section']
+  },
+  {
+    name: 'Real World',
+    icons: [
+      'eye',
+      'eye-slash',
+      'gift',
+      'trash',
+      'trash-alt',
+      'tag',
+      'award',
+      'briefcase',
+      'funnel',
+      'credit-card',
+      'sticky',
+      'book',
+      'light-bulb',
+      'light-bulb-off',
+      'megaphone',
+      'mortarboard',
+      'scissors',
+      'paper-clip',
+      'gem',
+      'hourglass',
+      'hourglass-bottom',
+      'hourglass-top',
+      'hourglass-split',
+      'pin',
+      'pin-angle',
+      'band-aid',
+      'ticket',
+      'box',
+      'pass',
+      'rocket',
+      'newspaper',
+      'note'
+    ]
+  },
+  {
+    name: 'Buildings',
+    icons: ['house', 'building', 'buildings', 'house-building', 'buildings-alt']
+  },
+  {
+    name: 'Emoji',
+    icons: ['emoji-smile', 'emoji-frown', 'emoji-neutral']
+  },
+  {
+    name: 'Hands',
+    icons: ['hand-index', 'hand-thumbs-up', 'hand-thumbs-down']
+  },
+  {
+    name: 'Social',
+    icons: [
+      'facebook',
+      'instagram',
+      'tiktok',
+      'linkedin',
+      'twitter',
+      'youtube',
+      'skype',
+      'whatsapp',
+      'spotify'
+    ]
+  },
+  {
+    name: 'Platforms',
+    icons: ['android', 'apple', 'google-play', 'microsoft']
+  }
+]