+++ CHANGELOG.md
@@ -1,3 +1,126 @@
+# [1.7.0-beta.13](https://github.com/occmundial/atomic-icons/compare/v1.7.0-beta.12...v1.7.0-beta.13) (2024-08-24)
+
+
+### Features
+
+* Add heart-filled icon ([4093559](https://github.com/occmundial/atomic-icons/commit/409355908bde655ae39841eedd2abecaad8faee1))
+
+# [1.7.0-beta.12](https://github.com/occmundial/atomic-icons/compare/v1.7.0-beta.11...v1.7.0-beta.12) (2024-08-20)
+
+
+### Bug Fixes
+
+* Fix malformed icons ([61d7e79](https://github.com/occmundial/atomic-icons/commit/61d7e79c8b50de2142fa78682f4c8ce954d893d1))
+
+# [1.7.0-beta.11](https://github.com/occmundial/atomic-icons/compare/v1.7.0-beta.10...v1.7.0-beta.11) (2024-08-15)
+
+
+### Bug Fixes
+
+* Fix chanelog ([9f0a017](https://github.com/occmundial/atomic-icons/commit/9f0a017a8ea662c1c8ecab2852916a215a4e3204))
+
+# [1.7.0-beta.10](https://github.com/occmundial/atomic-icons/compare/v1.7.0-beta.9...v1.7.0-beta.10) (2024-08-15)
+
+
+### Features
+
+* Change aistars icon ([4f2bb52](https://github.com/occmundial/atomic-icons/commit/4f2bb52e9ec564c554e70e8ae92a67aa65a364cb))
+* Generate font file and glyph map ([e5b18e2](https://github.com/occmundial/atomic-icons/commit/e5b18e2514d72485d6a3d699b11e79d991e05c3a))
+
+# [1.7.0-beta.10](https://github.com/occmundial/atomic-icons/compare/v1.7.0-beta.9...v1.7.0-beta.10) (2024-08-15)
+
+
+### Features
+
+* Change aistars icon ([4f2bb52](https://github.com/occmundial/atomic-icons/commit/4f2bb52e9ec564c554e70e8ae92a67aa65a364cb))
+* Generate font file and glyph map ([e5b18e2](https://github.com/occmundial/atomic-icons/commit/e5b18e2514d72485d6a3d699b11e79d991e05c3a))
+
+# [1.7.0-beta.9](https://github.com/occmundial/atomic-icons/compare/v1.7.0-beta.8...v1.7.0-beta.9) (2024-08-07)
+
+
+### Features
+
+* Add dollar-alt icon ([82aa98f](https://github.com/occmundial/atomic-icons/commit/82aa98f57bd10b9e532d5eaf135b4ddc10564f94))
+
+# [1.7.0-beta.8](https://github.com/occmundial/atomic-icons/compare/v1.7.0-beta.7...v1.7.0-beta.8) (2024-04-24)
+
+
+### Features
+
+* Add app store and platform icons ([260f5b7](https://github.com/occmundial/atomic-icons/commit/260f5b7accb82fae4e3b26ed4ce702d053b31768))
+
+# [1.7.0-beta.7](https://github.com/occmundial/atomic-icons/compare/v1.7.0-beta.6...v1.7.0-beta.7) (2024-04-24)
+
+
+### Features
+
+* Add the last missing icons ([ecd2c62](https://github.com/occmundial/atomic-icons/commit/ecd2c62befcd55b924fe3fcaffb2033c8cd68498))
+
+# [1.7.0-beta.6](https://github.com/occmundial/atomic-icons/compare/v1.7.0-beta.5...v1.7.0-beta.6) (2024-04-18)
+
+
+### Bug Fixes
+
+* Remove fill black in aistars icon elements ([b0cf997](https://github.com/occmundial/atomic-icons/commit/b0cf997b0a6b3ecc3770649e120269ba3ddee3d8))
+
+# [1.7.0-beta.5](https://github.com/occmundial/atomic-icons/compare/v1.7.0-beta.4...v1.7.0-beta.5) (2024-04-18)
+
+
+### Bug Fixes
+
+* Remove fill of aistars icon ([2cf17c4](https://github.com/occmundial/atomic-icons/commit/2cf17c4502056dfd3db0aa528a6c7b5ed6ccb207))
+
+# [1.7.0-beta.4](https://github.com/occmundial/atomic-icons/compare/v1.7.0-beta.3...v1.7.0-beta.4) (2024-04-18)
+
+
+### Bug Fixes
+
+* Remove fill property of tag icon ([e24823b](https://github.com/occmundial/atomic-icons/commit/e24823bca7be9fafbdcb7d26bd1470e4e1f93f75))
+
+# [1.7.0-beta.3](https://github.com/occmundial/atomic-icons/compare/v1.7.0-beta.2...v1.7.0-beta.3) (2024-04-11)
+
+
+### Bug Fixes
+
+* Change star icons names ([fa307ca](https://github.com/occmundial/atomic-icons/commit/fa307ca287f5e5ed912cce571b8b8d1ec495288b))
+
+# [1.7.0-beta.2](https://github.com/occmundial/atomic-icons/compare/v1.7.0-beta.1...v1.7.0-beta.2) (2024-04-11)
+
+
+### Bug Fixes
+
+* Change phone/mobile icons ([dcb5465](https://github.com/occmundial/atomic-icons/commit/dcb546591791587efefc886c870ae950706838b0))
+
+# [1.7.0-beta.1](https://github.com/occmundial/atomic-icons/compare/v1.6.0...v1.7.0-beta.1) (2024-04-11)
+
+
+### Features
+
+* Add arrows icons ([a5332fb](https://github.com/occmundial/atomic-icons/commit/a5332fbefe29499c70642718fe67a3fdb538dd91))
+* Add box arrow icons ([d2121db](https://github.com/occmundial/atomic-icons/commit/d2121db7a1228f8a44924ffcfa015b73544fa0b7))
+* Add buildings icons ([698e334](https://github.com/occmundial/atomic-icons/commit/698e334e7ef07968752caa23d609bb94ca83f61e))
+* Add chevron icons ([6b0fa28](https://github.com/occmundial/atomic-icons/commit/6b0fa286bd9187e873c4dfb1e922aa288f873df2))
+* Add commerce icons ([9372077](https://github.com/occmundial/atomic-icons/commit/937207748aefda6100a25cdde2943cb22c0942d7))
+* Add communication icons ([f924754](https://github.com/occmundial/atomic-icons/commit/f924754551f8d7c406f839281c48ac90b82ae3a4))
+* Add controls icons ([729e34b](https://github.com/occmundial/atomic-icons/commit/729e34bb181864abe4d9900992d34baedbf92366))
+* Add devices icons ([5328fa5](https://github.com/occmundial/atomic-icons/commit/5328fa5b2464dd73ab67df6a638fbf11bbf9f66d))
+* Add emoji icons ([ee3feb4](https://github.com/occmundial/atomic-icons/commit/ee3feb4e1ed97bb24ee5d737ce93da75bb2f4290))
+* Add files folders icons ([5c4c365](https://github.com/occmundial/atomic-icons/commit/5c4c365738a62028cfc90276c03cb315973ca1a6))
+* Add hands icons ([4717086](https://github.com/occmundial/atomic-icons/commit/47170865c9cc35d58b5ec2da9db859f1012db3f3))
+* Add layout icons ([6f00a99](https://github.com/occmundial/atomic-icons/commit/6f00a9948bd570bf4ddcb114890e4a6ed5083d84))
+* Add miscelaneous icons ([a39ff39](https://github.com/occmundial/atomic-icons/commit/a39ff3937a6e8dae68c92d5bb702db84c52ef225))
+* Add people icons ([e8c8141](https://github.com/occmundial/atomic-icons/commit/e8c8141329a905dd0482b03f2404c6ecab1670ef))
+* Add real world icons ([70cfbf3](https://github.com/occmundial/atomic-icons/commit/70cfbf3e920e350fe811aca3b040993e4b14f161))
+* Add security icons ([cbc14ca](https://github.com/occmundial/atomic-icons/commit/cbc14ca333c8d64cfc4c4733040a3f4e9de4be75))
+* Add social icons ([9d4749b](https://github.com/occmundial/atomic-icons/commit/9d4749b5fd197d66d754b572af8814d30b56d582))
+* Add sort and filter icons ([283fc55](https://github.com/occmundial/atomic-icons/commit/283fc550409b74e8b190b997937023ad887aef00))
+* Add symbols icons ([90a89eb](https://github.com/occmundial/atomic-icons/commit/90a89eb35e08c213b23a240c9fbd33e8acb8b14c))
+* Add time icons ([b1e473a](https://github.com/occmundial/atomic-icons/commit/b1e473afa6146a1f04a13319f8e0b790f76528a5))
+* Add tools icons ([e01bf0b](https://github.com/occmundial/atomic-icons/commit/e01bf0b30795f65714e83c82770c3e1d6af9c230))
+* Add typography icons ([8b38022](https://github.com/occmundial/atomic-icons/commit/8b380224da491671afc51f27270fb7793176a225))
+* Add weather icons ([3451376](https://github.com/occmundial/atomic-icons/commit/3451376c6afb46d37ca508a552bf8558e88e707a))
+* Upgrade atomic ([f22c3a9](https://github.com/occmundial/atomic-icons/commit/f22c3a9c59dd34b5d8e3f078e352122bbda26e40))
+
 # [1.6.0](https://github.com/occmundial/atomic-icons/compare/v1.5.1...v1.6.0) (2023-05-23)